Updated Header and Footer - Added Employee Page
I took the code from the header and the footer and put them in separate pages, I than used php with the include function for both.  That  way, if we change something in one header(I mean the div tage - top-part) it will affect all pages instead of one.   I also added the employee page and tested this out as well.